requirements.txt
Django==4.2
djangorestframework==3.14.0
asymmetric-jwt-auth==1.0.0